requirements.txt
streamlit
azure-cosmos
